Replace shell process by Solr process
Fixes #1

Before this fix the Solr process was spawned as a child process of a shell
process. When run with supervisord it was necessary to stop Solr as group (kill
the Solr and the shell process). However this can lead to problems when
restarting Solr because the Solr process may need more time to stop than the
shell process but supervisord is only checking the exit status of the parent
process when stopping a process group. This can lead to the situation that the
new Solr process is started while the old one is still running.

To avoid this issue, we simply replace the shell process by the Solr process.